TEST: Split the staking test into multiple modules
