+++ fendermint/testing/contract-test/tests/staking/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2022-2023 Protocol Labs
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+pub mod machine;
+pub mod state;
