+++ fendermint/testing/contract-test/tests/staking/state.rs
@@ -1,36 +1,16 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
-#![allow(unused)]
-//! State Machine Test for the Staking contracts.
-//!
-//! The test simulates random actions validators can take, such as depositing and withdrawing
-//! collateral, and executes these actions on the actual Solidity contracts as well as an
-//! idealised model, comparing the results and testing that invariants are maintained.
-//!
-//! It can be executed the following way:
-//!
-//! ```text
-//! cargo test --release -p contract-test --test staking
-//! ```
-use std::collections::{HashMap, VecDeque};
-use std::sync::Arc;
-use std::time::Instant;
 
-use arbitrary::{Arbitrary, Unstructured};
+use std::collections::{HashMap, VecDeque};
+
+use arbitrary::Unstructured;
 use fendermint_crypto::{PublicKey, SecretKey};
 use fendermint_testing::arb::{ArbSubnetAddress, ArbSubnetID, ArbTokenAmount};
-use fendermint_testing::state_machine_seed;
-use fendermint_testing::{smt::StateMachine, state_machine_test};
 use fendermint_vm_core::Timestamp;
 use fendermint_vm_genesis::ipc::{GatewayParams, IpcParams};
 use fendermint_vm_genesis::{
-    Account, Actor, ActorMeta, Collateral, Genesis, Power, SignerAddr, Validator, ValidatorKey,
+    Account, Actor, ActorMeta, Collateral, Genesis, SignerAddr, Validator, ValidatorKey,
 };
-use fendermint_vm_interpreter::fvm::{
-    state::{ipc::GatewayCaller, FvmExecState},
-    store::memory::MemoryBlockstore,
-};
-use fvm::engine::MultiEngine;
 use fvm_shared::address::Address;
 use fvm_shared::bigint::BigInt;
 use fvm_shared::bigint::Integer;
@@ -39,58 +19,50 @@ use ipc_sdk::subnet_id::SubnetID;
 use rand::rngs::StdRng;
 use rand::SeedableRng;
 
-/// System Under Test for staking.
-struct StakingSystem {
-    /// FVM state initialized with the parent genesis, and a subnet created for the child.
-    parent_state: FvmExecState<MemoryBlockstore>,
-    /// Facilitate calling the gateway.
-    gateway: GatewayCaller<MemoryBlockstore>,
-}
-
 #[derive(Debug, Clone)]
-enum StakingOp {
+pub enum StakingOp {
     Deposit(TokenAmount),
     Withdraw(TokenAmount),
 }
 
 /// The staking message that goes towards the subnet to increase or decrease power.
 #[derive(Debug, Clone)]
-struct StakingUpdate {
-    configuration_number: u64,
-    addr: Address,
-    op: StakingOp,
+pub struct StakingUpdate {
+    pub configuration_number: u64,
+    pub addr: Address,
+    pub op: StakingOp,
 }
 
 #[derive(Debug, Clone)]
-struct StakingAccount {
-    public_key: PublicKey,
-    secret_key: SecretKey,
-    addr: Address,
+pub struct StakingAccount {
+    pub public_key: PublicKey,
+    pub secret_key: SecretKey,
+    pub addr: Address,
     /// In this test the accounts should never gain more than their initial balance.
-    initial_balance: TokenAmount,
+    pub initial_balance: TokenAmount,
     /// Initial stake this account is going to put into the subnet.
-    initial_stake: TokenAmount,
+    pub initial_stake: TokenAmount,
     /// Balance after the effects of deposits/withdrawals.
-    current_balance: TokenAmount,
+    pub current_balance: TokenAmount,
 }
 
 /// Reference implementation for staking.
 #[derive(Debug, Clone)]
-struct StakingState {
+pub struct StakingState {
     /// Accounts with secret key of accounts in case the contract wants to validate signatures.
-    accounts: HashMap<Address, StakingAccount>,
+    pub accounts: HashMap<Address, StakingAccount>,
     /// List of account addresses to help pick a random one.
-    addrs: Vec<Address>,
+    pub addrs: Vec<Address>,
     /// The parent genesis should include a bunch of accounts we can use to join a subnet.
-    parent_genesis: Genesis,
+    pub parent_genesis: Genesis,
     /// The child genesis describes the initial validator set to join the subnet
-    child_genesis: Genesis,
+    pub child_genesis: Genesis,
     /// Currently active child validator set.
-    child_validators: HashMap<Address, Collateral>,
+    pub child_validators: HashMap<Address, Collateral>,
     /// The configuration number to be incremented before each staking operation; 0 belongs to the genesis.
-    configuration_number: u64,
+    pub configuration_number: u64,
     /// Unconfirmed staking operations.
-    pending_updates: VecDeque<StakingUpdate>,
+    pub pending_updates: VecDeque<StakingUpdate>,
 }
 
 impl StakingState {
@@ -138,7 +110,7 @@ impl StakingState {
             let update = self.pending_updates.pop_front().expect("checked non-empty");
             match update.op {
                 StakingOp::Deposit(v) => {
-                    let mut power = self.child_validators.entry(update.addr).or_default();
+                    let power = self.child_validators.entry(update.addr).or_default();
                     power.0 += v;
                 }
                 StakingOp::Withdraw(v) => {
@@ -153,7 +125,7 @@ impl StakingState {
                                 e.insert(Collateral(c - v.clone()));
                             }
 
-                            let mut a = self
+                            let a = self
                                 .accounts
                                 .get_mut(&update.addr)
                                 .expect("validators have accounts");
@@ -175,7 +147,7 @@ impl StakingState {
     pub fn stake(mut self, addr: Address, value: TokenAmount) -> Self {
         self.configuration_number += 1;
 
-        let mut a = self.accounts.get_mut(&addr).expect("accounts exist");
+        let a = self.accounts.get_mut(&addr).expect("accounts exist");
 
         // Sanity check that we are generating the expected kind of values.
         // Using `debug_assert!` on the reference state to differentiate from assertions on the SUT.
@@ -342,149 +314,3 @@ impl arbitrary::Arbitrary<'_> for StakingState {
         Ok(StakingState::new(accounts, parent_genesis, child_genesis))
     }
 }
-
-enum StakingCommand {
-    /// Bottom-up checkpoint; confirms all staking operations up to the configuration number.
-    Checkpoint { next_configuration_number: u64 },
-    /// Increase the collateral of a validator; when it goes from 0 this means joining the subnet.
-    Stake(Address, TokenAmount),
-    /// Decrease the collateral of a validator; if it goes to 0 it means leaving the subnet.
-    Unstake(Address, TokenAmount),
-}
-
-#[derive(Default)]
-struct StakingMachine {
-    multi_engine: Arc<MultiEngine>,
-}
-
-impl StateMachine for StakingMachine {
-    type System = StakingSystem;
-
-    type State = StakingState;
-
-    type Command = StakingCommand;
-
-    type Result = ();
-
-    fn gen_state(&self, u: &mut Unstructured) -> arbitrary::Result<Self::State> {
-        StakingState::arbitrary(u)
-    }
-
-    fn new_system(&self, state: &Self::State) -> Self::System {
-        let rt = tokio::runtime::Runtime::new().expect("create tokio runtime for init");
-
-        let (parent_state, _) = rt
-            .block_on(contract_test::init_exec_state(
-                self.multi_engine.clone(),
-                state.parent_genesis.clone(),
-            ))
-            .expect("failed to init parent");
-
-        let gateway = GatewayCaller::default();
-
-        // TODO: Call the methods on the gateway to establish the subnet based on `state.child_genesis`:
-        // * Create the subnet with the given ID
-        // * Make all the validators join the subnet by putting down collateral according to their power
-
-        StakingSystem {
-            parent_state,
-            gateway,
-        }
-    }
-
-    fn gen_command(
-        &self,
-        u: &mut Unstructured,
-        state: &Self::State,
-    ) -> arbitrary::Result<Self::Command> {
-        let cmd = match u.choose(&["checkpoint", "stake", "unstake"]).unwrap() {
-            &"checkpoint" => {
-                let cn = match state.pending_updates.len() {
-                    0 => state.configuration_number,
-                    n => {
-                        let idx = u.choose_index(n).expect("non-zero");
-                        state.pending_updates[idx].configuration_number
-                    }
-                };
-                StakingCommand::Checkpoint {
-                    next_configuration_number: cn,
-                }
-            }
-            &"stake" => {
-                let a = u.choose(&state.addrs).expect("accounts not empty");
-                let a = state.accounts.get(a).expect("account exists");
-                // Limit ourselves to the outstanding balance - the user would not be able to send more value to the contract.
-                let b = BigInt::arbitrary(u)?.mod_floor(a.current_balance.atto());
-                let b = TokenAmount::from_atto(b);
-                StakingCommand::Stake(a.addr, b)
-            }
-            &"unstake" => {
-                let a = u.choose(&state.addrs).expect("accounts not empty");
-                let a = state.accounts.get(a).expect("account exists");
-                // We can try sending requests to unbond arbitrarily large amounts of collateral - the system should catch any attempt to steal.
-                // Only limiting it to be under the initial balance so that it's comparable to what the deposits could have been.
-                let b = BigInt::arbitrary(u)?.mod_floor(a.initial_balance.atto());
-                let b = TokenAmount::from_atto(b);
-                StakingCommand::Unstake(a.addr, b)
-            }
-            other => unimplemented!("unknown command: {other}"),
-        };
-        Ok(cmd)
-    }
-
-    fn run_command(&self, system: &mut Self::System, cmd: &Self::Command) -> Self::Result {
-        // TODO: Execute the command against the contract.
-    }
-
-    fn check_result(&self, cmd: &Self::Command, pre_state: &Self::State, result: &Self::Result) {
-        // TODO: Check that events emitted by the system are as expected.
-    }
-
-    fn next_state(&self, cmd: &Self::Command, state: Self::State) -> Self::State {
-        match cmd {
-            StakingCommand::Checkpoint {
-                next_configuration_number,
-            } => state.checkpoint(*next_configuration_number),
-            StakingCommand::Stake(addr, value) => state.stake(*addr, value.clone()),
-            StakingCommand::Unstake(addr, value) => state.unstake(*addr, value.clone()),
-        }
-    }
-
-    fn check_system(
-        &self,
-        cmd: &Self::Command,
-        post_state: &Self::State,
-        post_system: &Self::System,
-    ) {
-        match cmd {
-            StakingCommand::Checkpoint { .. } => {
-                // Sanity check the reference state while we have no contract to compare with.
-                debug_assert!(
-                    post_state
-                        .accounts
-                        .iter()
-                        .all(|(_, a)| a.current_balance <= a.initial_balance),
-                    "no account goes over initial balance"
-                );
-
-                debug_assert!(
-                    post_state
-                        .child_validators
-                        .iter()
-                        .all(|(_, p)| !p.0.is_zero()),
-                    "all child validators have non-zero collateral"
-                );
-            }
-            StakingCommand::Stake(addr, _) | StakingCommand::Unstake(addr, _) => {
-                let a = post_state.accounts.get(addr).unwrap();
-                debug_assert!(a.current_balance <= a.initial_balance);
-            }
-        }
-
-        // TODO: Compare the system with the state:
-        // * check that balances match
-        // * check that active powers match
-    }
-}
-
-state_machine_test!(staking, 20000 ms, 100 steps, StakingMachine::default());
